requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9  # For Alembic migrations
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0

# Vector Database
qdrant-client==1.7.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# LLM integration
openai==1.3.7
tiktoken==0.5.2

# Embeddings
numpy==1.24.3
scikit-learn==1.3.2

# Async and utilities
httpx==0.25.2
python-json-logger==2.0.7
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1